+++ BE/models/Message.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, String, Text, DateTime, func, ForeignKey, Enum as SQLEnum
+from sqlalchemy import Column, Integer, String, Text, DateTime, func, ForeignKey, Enum as SQLEnum
 from sqlalchemy.orm import relationship
 from BE.db.session import Base
 import enum
@@ -10,8 +10,8 @@ class MessageType(enum.Enum):
 class Message(Base):
     __tablename__ = "tblMessage"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    chat_id = Column(BigInteger, ForeignKey("tblChat.id"), nullable=False, index=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    chat_id = Column(Integer, ForeignKey("tblChat.id"), nullable=False, index=True)
     type = Column(SQLEnum(MessageType), nullable=False)
     content = Column(Text, nullable=False)
     created_at = Column(DateTime, server_default=func.now())
